applied patch from Aurimas Juška that adds a missing cast which should
2006-10-10  Sven Neumann  <sven@gimp.org>

	* plug-ins/common/nova.c: applied patch from Aurimas Juška that
	adds a missing cast which should fix bug #340073.

+++ plug-ins/common/nova.c
@@ -799,9 +799,11 @@ nova (GimpDrawable *drawable,
            for (col = 0, x = 0; col < x2; col++, x++)
              {
                u = (gdouble) (x - xc) /
-                          (pvals.radius * width / drawable->width * zoom);
+                          ((gdouble) pvals.radius * width /
+                              drawable->width * zoom);
                v = (gdouble) (y - yc) /
-                          (pvals.radius * height / drawable->height * zoom);
+                          ((gdouble) pvals.radius * height /
+                              drawable->height * zoom);
                l = sqrt (u * u + v * v);
 
                /* This algorithm is still under construction. */
